Rename `general` module to `streaming`
