+++ api/contextqa/services/chat.py
@@ -18,8 +18,8 @@ from contextqa import settings
 from contextqa.agents.tools import searcher
 from contextqa.models.schemas import LLMQueryRequest
 from contextqa.utils import memory, prompts
-from contextqa.utils.general import stream
 from contextqa.agents.tools import searcher
+from contextqa.utils.streaming import stream
 
 
 _MESSAGES = [
